dev sync

+++ Client/trunk/ParaEngineClient/3dengine/ParaEngineSettings.h
@@ -217,7 +217,7 @@ namespace ParaEngine
 		ATTRIBUTE_METHOD1(ParaEngineSettings, GetWritablePath_s, const char**) { *p1 = cls->GetWritablePath(); return S_OK; }
 		ATTRIBUTE_METHOD1(ParaEngineSettings, SetWritablePath_s, const char*) { cls->SetWritablePath(p1); return S_OK; }
 
-		ATTRIBUTE_METHOD1(ParaEngineSettings, GetAppHWND_s, int*) { *p1 = cls->GetAppHWND(); return S_OK; }
+		ATTRIBUTE_METHOD1(ParaEngineSettings, GetAppHWND_s, double*) { *p1 = (double)cls->GetAppHWND(); return S_OK; }
 	public:
 		/** if true, IO is restricted to current working directory and writable directory.  Under win32, this is true by default. */
 		static bool IsSandboxMode();
@@ -690,7 +690,7 @@ namespace ParaEngine
 		*/
 		void SetWritablePath(const char* sPath);
 
-		static int GetAppHWND();
+		static intptr_t GetAppHWND();
 	protected:
 		void LoadNameIndex();
 	private:
